feat: host/guest flow improvements, geocoded search, media upload and supabase migrations
- add host reservations recap, QR scan and check-in confirmation screens

- add new/edit listing flows with map picker and unified address suggestions

- implement Mapbox/Nominatim geocoding helper and distance-based search filtering

- improve profile editing, host/guest mode switch UX, phone fields and avatar upload behavior

- fix web picker behavior and multiple react-native-web deprecation warnings

- update Supabase SQL scripts/migrations/seeds for RLS, phone columns, decimal prices and storage policies

+++ supabase/migrations/20260304200100_schema_base.sql
@@ -4,7 +4,7 @@ create table if not exists public.services (
   id uuid primary key default gen_random_uuid(),
   title text not null,
   category text not null check (category in ('rest', 'shower', 'storage')),
-  price_eur numeric(10,2) not null,
+  price_eur integer not null,
   location text not null,
   city text,
   region text,
